hiC_query.py modified to support searching of HiC tables of multiple replicates in multiple cell lines automatically.

+++ hiC_query.py
@@ -86,35 +86,45 @@ def process_inputs(inputs):
 			interactions[locus].append(interaction)
 	return interactions"""
 
-def find_interactions(snps,db_dir,distance):
+def find_interactions(snps,db_connection_limit,distance):
 	print "Finding interactions..."
 	interactions = {}
 	open_dbs = []
-	db_dir = db_dir.strip('/') #Strip trailing '/' if it is there
-	table_index_db = sqlite3.connect(db_dir + "/dummy.db")
-	table_index_db.text_factory = str
-	table_index = table_index_db.cursor()
-	for snp in snps.keys():
-		print "\tFinding interactions for " + str(snp)
-		snp_chr = snps[snp][0]
-		snp_locus = snps[snp][1]
-		if snp_chr not in open_dbs:
-			if len(open_dbs) >= 10:
-				table_index.execute("DETACH db" + open_dbs.pop())
-			table_index.execute("ATTACH '" + db_dir + '/' + db_dir[db_dir.rfind('/'):db_dir.rfind('_')] + '_' + snp_chr + ".db' AS db" + snp_chr)
-		else:
-			open_dbs.remove(snp_chr)
-		open_dbs.insert(0,snp_chr)
-		interactions[snp] = []
-		for interaction in table_index.execute("SELECT chr1, locus1 FROM db" + snp_chr + ".interactions WHERE chr2=? and (locus2 >= ? and locus2 <= ?)",[snp_chr,snp_locus-distance,snp_locus+distance]):
-			interactions[snp].append(interaction)
-		for interaction in table_index.execute("SELECT chr2, locus2 FROM db" + snp_chr + ".interactions WHERE chr1=? and (locus1 >= ? and locus1 <= ?)",[snp_chr,snp_locus-distance,snp_locus+distance]):
-			interactions[snp].append(interaction)
-	#outfile = open("db_index_output_test.txt",'w')
-	#for snp in interactions.keys():
-	#	for interaction in interactions[snp]:
-	#		outfile.write(snp_chr + '\t' + str(snp_locus) + '\t' + interaction[0] + '\t' + str(interaction[1]) + '\n')
-	os.remove(db_dir + "/dummy.db")
+	#db_dir = db_dir.strip('/') #Strip trailing '/' if it is there
+	for cell_line in os.listdir("hic_data"):
+		if os.path.isdir("hic_data/" + cell_line):
+			print "\tSearching cell line " + cell_line
+			for replicate in os.listdir("hic_data/" + cell_line):
+				db_dir = "hic_data/" + cell_line + '/' + replicate
+				if os.path.isdir(db_dir):
+					if all_chrs_present(db_dir):
+						print "\t\tSearching replicate " + replicate
+						table_index_db = sqlite3.connect(db_dir + "/dummy.db")
+						table_index_db.text_factory = str
+						table_index = table_index_db.cursor()
+						for snp in snps.keys():
+							print "\t\t\tFinding interactions for " + str(snp)
+							snp_chr = snps[snp][0]
+							snp_locus = snps[snp][1]
+							if snp_chr not in open_dbs:
+								if len(open_dbs) >= db_connection_limit:
+									table_index.execute("DETACH db" + open_dbs.pop())
+								table_index.execute("ATTACH '" + db_dir + '/' + replicate + '_' + snp_chr + ".db' AS db" + snp_chr)
+							else:
+								open_dbs.remove(snp_chr)
+							open_dbs.insert(0,snp_chr)
+							interactions[snp] = []
+							for interaction in table_index.execute("SELECT chr1, locus1 FROM db" + snp_chr + ".interactions WHERE chr2=? and (locus2 >= ? and locus2 <= ?)",[snp_chr,snp_locus-distance,snp_locus+distance]):
+								interactions[snp].append(interaction)
+							for interaction in table_index.execute("SELECT chr2, locus2 FROM db" + snp_chr + ".interactions WHERE chr1=? and (locus1 >= ? and locus1 <= ?)",[snp_chr,snp_locus-distance,snp_locus+distance]):
+								interactions[snp].append(interaction)
+						#outfile = open("db_index_output_test.txt",'w')
+						#for snp in interactions.keys():
+						#	for interaction in interactions[snp]:
+						#		outfile.write(snp_chr + '\t' + str(snp_locus) + '\t' + interaction[0] + '\t' + str(interaction[1]) + '\n')
+						os.remove(db_dir + "/dummy.db")
+					else:
+						print "\t\tWarning: database for replicate %s not in expected format." % replicate
 	return interactions
 
 """def find_interactions_raw(loci,table_fp,distance,mapq_cutoff): ##Read raw HiC file, no indexing
@@ -149,7 +159,7 @@ def find_eqtls(interactions):
 	for snp in interactions.keys():
 		eqtls[snp] = {}
 	for db in os.listdir("eQTLs"): #Iterate through databases of eQTLs by tissue type
-		print "\tQuerying " + db
+		print "\tQuerying " + db[:db.rfind('.')]
 		eqtl_index_db = sqlite3.connect("eQTLs/" + db)
 		eqtl_index_db.text_factory = str
 		eqtl_index = eqtl_index_db.cursor()
@@ -162,22 +172,25 @@ def find_eqtls(interactions):
 						eqtls[snp][tissue].append(eqtl)
 	return eqtls
 
+def all_chrs_present(db_dir):
+	chr_set = Set(["1","2","3","4","5","6","7","8","9","10","11","12","13","14","15","16","17","18","19","20","21","22","X","Y","MT"])
+	check_set = Set([])
+	for db in os.listdir(db_dir): #Perform check to see if replicate database is in expected configuration, i.e. 25 separate databases, one for each chromosome
+		check_set.add(db[db.rfind('_')+1:db.rfind('.')])
+	return check_set == chr_set
+
 
 if __name__ == "__main__":
 	parser = argparse.ArgumentParser(description="")
 	parser.add_argument("-i","--inputs",nargs='+',help="The the dbSNP IDs or loci of SNPs of interest in the format \"chr<x>:<locus>\"")
-	parser.add_argument("-t","--table_fp",default="HIC059_mapq150_chr21.test.txt",help="TEMP: The Rao table from which to pull connections.")
-	parser.add_argument("-x","--db_dir",default="GSM1551552_HIC003_merged_nodups_db",help="TEMP: The directory containing the chromosome-indexed database to search.")
+	#parser.add_argument("-t","--table_fp",default="HIC059_mapq150_chr21.test.txt",help="TEMP: The Rao table from which to pull connections.")
+	#parser.add_argument("-x","--db_dir",default="GSM1551552_HIC003_merged_nodups_db",help="TEMP: The directory containing the chromosome-indexed database to search.")
 	parser.add_argument("-d","--distance",type=int,default=500,help="The allowed distance from the locus of interest for a fragment to be considered.")
+	parser.add_argument("-c","--db_connection_limit",type=int,default=4)
 	#parser.add_argument("-m","--mapq_cutoff",type=int,default=150,help="The minimum mapq score allowed.")
 	args = parser.parse_args()
-	inputs = args.inputs
-	table_fp = args.table_fp
-	db_dir = args.db_dir
-	distance = args.distance
-	#mapq_cutoff = args.mapq_cutoff
-	snps = process_inputs(inputs)
-	interactions = find_interactions(snps,db_dir,distance)
+	snps = process_inputs(args.inputs)
+	interactions = find_interactions(snps,args.db_connection_limit,args.distance)
 	eqtls = find_eqtls(interactions)
 	for snp in interactions.keys():
 		print snp + ':'
